refactor(FE-6): edit components folder, add scss, browser router dom and also absolute import
