+++ my-app/src/App.js
@@ -1,14 +1,17 @@
-import Discount from "./components/Discount/Discount";
-import Header from "./components/Header/Header";
-import SignUp from "./components/SignUp/SignUp";
+import React from "react";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+
+import Home from "pages/Home/Home";
 
 function App() {
   return (
-    <div>
-      <Header />
-      <Discount />
-      <SignUp />
-    </div>
+    <BrowserRouter>
+      <div>
+        <Routes>
+          <Route path="/" element={<Home />} />
+        </Routes>
+      </div>
+    </BrowserRouter>
   );
 }
