+++ my-app/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import styles from './index.scss'
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
