+++ my-app/src/shared/ui/Discount/styles.module.scss
